Rename to DatasetReader, implement factory method
